format edits & added lessons learned & added tech applicability
added my parts in conclusion : lessons learned and technological applicably NOTE that i changed the order to make it more relateable when reading, feel free to edit if u have another opinion. ----- added the AI stuff at the appendix. -----Edit the format for non-functional requirements. and some of the content on it from my parts.